🚚 refactor: 혼합 배치된 순수 유틸/타입을 하위 레이어로 분리 (옵션 2)
[변경 내용]
- 순수 유틸을 shared로 이동:
  - loginRedirect(로그인 리다이렉트 URL 로직) auth/lib -> shared/lib
  - roleTagMapper(toRoleTag, 이전 auth/model/mappers) -> shared/lib
  - LoginCircleButton(범용 소셜 버튼 UI) login/ui -> shared/ui
- storage(범용 파일 업로드 API)를 project/new/api -> shared/api로 분리. 업로드 계약
  타입(PrepareUpload*/UploadCategory) 3개도 project types.ts에서 storage로 인라인
- terms 도메인 타입(Term, TermType)을 auth/model/types -> shared/model/domain,
  auth는 re-export로 하위 호환 유지
- 소비자 import 경로 갱신, 래칫 baseline 20 -> 15

[변경 이유]
- 이 요소들은 특정 feature에 있지만 도메인 의존이 없는 순수 유틸/타입이라 다른
  feature가 참조하며 features->features 위반을 유발. 올바른 위치는 하위 레이어
  - loginRedirect/roleTagMapper: 순수 로직 -> shared/lib
  - LoginCircleButton: 도메인 모르는 범용 UI -> shared/ui
  - storage: 프로젝트와 무관한 파일 업로드 -> shared/api
  - Term/TermType: signup과 auth가 공유하는 약관 도메인 enum -> shared/model
- boundaries 위반 20 -> 15건 (-5): settings->auth 2, settings->project 1,
  settings->login 1, signup->auth 1 해소. 동작 변경 없음(순수 이동)

[고려 사항]
- storage 이동 시 project/new/api/types.ts의 업로드 타입 3개를 분리하고
  project/new/api/index.ts 배럴은 shared에서 re-export해 기존 소비자 호환 유지
- auth/model/types는 Term/TermType을 shared에서 import(로컬 사용) + re-export(외부
  호환) 둘 다 유지
- useSchools/api/school -> organization은 보류: auth판 SchoolNameListResponse가
  org판과 다른 타입(schoolId number vs string)이라 rename/충돌 정리 비용이 이득
  (1건) 대비 큼. 별도 판단 필요
- 남은 15건은 clean 파일 분리 영역 밖: settings의 계정 관리가 auth 기능(이메일
  인증/credentials/OAuth 연동) 조합, application/matching의 프로젝트 도메인 UI
  표시 등. auth 기능의 하위 레이어 추출 또는 페이지 레벨 재조립(동작 검증 필요)이
  필요한 별도 영역
- ProjectApplyModal.ui.test 5건 실패는 기존 이슈로 무관
- 런타임 검증: tsc + 프로덕션 빌드 성공
