+++ src/app/authWiring.ts
@@ -1,12 +1,12 @@
 import { useAuthStore } from "@/entities/member/store/authStore"
+import { setAuthBridge } from "@/shared/lib/authBridge"
 import {
   buildLoginRedirectSearch,
   getCurrentReturnTo,
-} from "@/features/auth/lib/loginRedirect"
-import { setAuthBridge } from "@/shared/lib/authBridge"
+} from "@/shared/lib/loginRedirect"
 
 // axios(shared)가 필요로 하는 인증 동작의 실제 구현을 조립 시점에 주입한다.
-// 도메인 store(entities)와 로그인 리다이렉트(features)를 아는 곳은 app 레이어뿐이다.
+// 도메인 store(entities/member)를 shared에 노출하지 않기 위해 app에서 주입한다.
 export function wireAuthBridge(): void {
   setAuthBridge({
     getAccessToken: () => useAuthStore.getState().accessToken,
