+++ src/features/auth/lib/logout.ts
@@ -1,5 +1,5 @@
 import { useAuthStore } from "@/entities/member/store/authStore"
-import { clearLoginReturnTo } from "@/features/auth/lib/loginRedirect"
+import { clearLoginReturnTo } from "@/shared/lib/loginRedirect"
 
 export function logout() {
   clearLoginReturnTo()
