+++ src/features/auth/model/types.ts
@@ -90,7 +90,7 @@ export interface RegisterResponse {
   refreshToken: string
 }
 
-export type TermType = "SERVICE" | "PRIVACY" | "MARKETING" | "LOCATION"
+export type { TermType } from "@/shared/model/domain"
 
 export interface TermResponse {
   id: number
@@ -98,16 +98,9 @@ export interface TermResponse {
   isMandatory: boolean
 }
 
-export interface Term {
-  id: number
-  type: TermType
-  typeDescription: string
-  link: string
-  isMandatory: boolean
-  version: number
-  createdAt: string
-  updatedAt: string
-}
+import type { Term } from "@/shared/model/domain"
+
+export type { Term } from "@/shared/model/domain"
 
 export interface Terms {
   terms: Term[]
