+++ src/routes/login/default.tsx
@@ -19,11 +19,6 @@ import {
 import { handleLoginResponse } from "@/features/auth/lib/handleLoginResponse"
 import { startKakaoSignIn } from "@/features/auth/lib/kakaoSignIn"
 import {
-  normalizeReturnTo,
-  rememberLoginReturnTo,
-  resolveLoginSuccessPath,
-} from "@/features/auth/lib/loginRedirect"
-import {
   Divider,
   Input,
   LoginCircleButton,
@@ -32,6 +27,11 @@ import {
 } from "@/features/login"
 import { emailSchema } from "@/features/signup/validation"
 import CheckIcon from "@/shared/assets/icon/check/CheckIcon"
+import {
+  normalizeReturnTo,
+  rememberLoginReturnTo,
+  resolveLoginSuccessPath,
+} from "@/shared/lib/loginRedirect"
 import { createMeta, SITE_URL } from "@/shared/seo"
 import { Button } from "@/shared/ui/Button"
 import { TextButton } from "@/shared/ui/button/TextButton"
