+++ src/widgets/navigation/header/ProfileDropdown.tsx
@@ -6,9 +6,9 @@ import { isCentralStaff, isSuperAdmin } from "@/entities/member/model/identity"
 import { useAuthStore } from "@/entities/member/store/authStore"
 import { logout as apiLogout } from "@/features/auth/api/credentials"
 import { logout as localLogout } from "@/features/auth/lib/logout"
-import { toRoleTag } from "@/features/auth/model/mappers"
 import { useSelectedChallengerStore } from "@/features/auth/store/selectedChallengerStore"
 import ProfileIcon from "@/shared/assets/icon/people/ProfileIcon"
+import { toRoleTag } from "@/shared/lib/roleTagMapper"
 import { cn } from "@/shared/lib/utils"
 import { TextButton } from "@/shared/ui/button/TextButton"
 import { RoleTagChip } from "@/shared/ui/chip/RoleTagChip"
